examples/github_coding_agent/requirements.txt
argparse
requests
python-dotenv
pydantic
llama-stack-client
llama-models
# Required for llama-models
blobfile